messed with styling a little more and moved some files into folders to organize better.
