+++ outreach-mapping-solutions-fe-react/src/clientPortal.css
@@ -1,0 +1,3 @@
+.client-portal-container {
+  display: flex;
+}
