+++ outreach-mapping-solutions-fe-react/src/clientPortal.js
@@ -1,11 +1,11 @@
-import { ClientBaseForm } from "./clientBaseForm";
-import { ClientSearchResults } from "./clientSearchResults";
+import { ClientBaseForm } from "./clientBaseForm/clientBaseForm";
+import { ClientSearchResults } from "./clientSearchResults/clientSearchResults";
 
 export function ClientPortal() {
   return (
-    <>
+    <div className="client-portal-container">
       <ClientBaseForm />
       <ClientSearchResults />
-    </>
+    </div>
   );
 }
