+++ outreach-mapping-solutions-fe-react/src/index.css
@@ -1,5 +1,19 @@
 .portal-main-container {
   display: flex;
   flex-direction: row;
-  justify-content: space-around;
+  /* border-style: solid; */
+  gap: 20px;
+}
+
+.nav-bar-main {
+  /* border-style: solid; */
+  justify-content: center;
+  padding: 10px;
+  border: 1;
+}
+
+.selected-portal {
+  width: 60%;
+  justify-content: left;
+  /* border-style: solid; */
 }
