--- outreach-mapping-solutions-fe-react/src/monthDaySelector.js
@@ -1,25 +1,0 @@
-// prettier-ignore
-const days29=['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24', '25', '26', '27', '28', '29'];
-// prettier-ignore
-const days30=['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24', '25', '26', '27', '28', '29', '30'];
-// prettier-ignore
-const days31=['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24', '25', '26', '27', '28', '29', '30', '31'];
-
-//prettier-ignore
-const months31Days=["January", "March", "May", "July", "August", "October", "December"];
-const months30Days = ["April", "June", "September", "November"];
-
-const monthsDays = [
-  { month: "January", days: "30" },
-  { month: "February", days: "29" },
-  { month: "March", days: "31" },
-  { month: "April", days: "30" },
-  { month: "May", days: "31" },
-  { month: "June", days: "30" },
-  { month: "July", days: "31" },
-  { month: "August", days: "31" },
-  { month: "September", days: "30" },
-  { month: "October", days: "31" },
-  { month: "November", days: "30" },
-  { month: "December", days: "31" },
-];
