+++ outreach-mapping-solutions-fe-react/src/navBarMain/navBarMain.css
@@ -1,6 +1,6 @@
 .nav-bar-main {
   display: flex;
-  justify-content: space-between;
+  /* justify-content: space-between; */
 }
 
 .nav-bar-links {
